# examples/minimal/internal/transport/http/notifications.go
// Package http provides HTTP handlers for the service.
// ============================================================================
// AUTO-GENERATED by ANG (Architectural Normalized Generator)
// Version: 0.1.127
// InputHash: 0042c83541f19460bd5eeaaca878da6acf4e7b35b891888724c9c4cdb759de2b
// CompilerHash: 294b50c17c5868adf98ff6e7e34d9b0ddcf67156d863f2f508ad57df44889164
// DO NOT EDIT - changes will be overwritten on next generation
// GENERATED: true
// ============================================================================

package http

import (
	"encoding/json"
	"net/http"

	"github.com/example/minimal/internal/pkg/errors"
	"github.com/example/minimal/internal/port"
	"github.com/go-chi/chi/v5"
)

func RegisterNotificationsRoutes(r chi.Router, svc port.Notifications) {
	r.With(MaxBodySizeMiddleware(1048576)).Post("/notifications/email", handlerSendNoticeEmail(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Post("/notifications/invitation", handlerSendInvitationEmail(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Post("/notifications/password-reset", handlerSendPasswordResetEmail(svc))
}

// handlerSendNoticeEmail
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerSendNoticeEmail(svc port.Notifications) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.SendNoticeEmailRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.SendNoticeEmail(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerSendInvitationEmail
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerSendInvitationEmail(svc port.Notifications) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.SendInvitationEmailRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.SendInvitationEmail(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerSendPasswordResetEmail
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerSendPasswordResetEmail(svc port.Notifications) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.SendPasswordResetEmailRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.SendPasswordResetEmail(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}