# python/api/openapi.yaml
openapi: 3.0.3
info:
  title: ANG API
  version: 0.1.0
  x-ang-version: 0.1.127
  x-ang-input-hash: 7d9ba8054e7ed584000a9c7ea12959fe1c66741726ec018e6c1e6f79ce78470c
  x-ang-compiler-hash: 294b50c17c5868adf98ff6e7e34d9b0ddcf67156d863f2f508ad57df44889164
paths:
...
  /auth/login:
    post:
      summary: Login
      operationId: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /auth/profile:
    put:
      summary: UpdateProfile
      operationId: UpdateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /auth/register:
    post:
      summary: Register
      operationId: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /notifications/invitation:
    post:
      summary: SendInvitationEmail
      operationId: SendInvitationEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationEmailRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationEmailResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /notifications/password-reset:
    post:
      summary: SendPasswordResetEmail
      operationId: SendPasswordResetEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPasswordResetEmailRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPasswordResetEmailResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: integer
        errors:
          type: object
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        accesstoken:
          type: string
        refreshtoken:
          type: string
        user:
          type: string
      required:
        - accesstoken
        - refreshtoken
        - user
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 64
        name:
          type: string
          minLength: 2
          maxLength: 100
      required:
        - email
        - password
        - name
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name
    SendInvitationEmailRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        invitername:
          type: string
          minLength: 2
          maxLength: 120
        inviteurl:
          type: string
          minLength: 12
          maxLength: 500
      required:
        - email
        - invitername
        - inviteurl
    SendInvitationEmailResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    SendPasswordResetEmailRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        reseturl:
          type: string
          minLength: 12
          maxLength: 500
      required:
        - email
        - reseturl
    SendPasswordResetEmailResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    UpdateProfileRequest:
      type: object
      properties:
        userid:
          type: string
        name:
          type: string
          minLength: 2
          maxLength: 100
        avatarurl:
          type: string
      required:
        - userid
    UpdateProfileResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT